Team olly
This is a weekend team event for students to showcase and develop web design & user experience skills. We have 10 spots open for WEB DEVELOPERS, sign up here if you can bring coding skills!

If you've attended a startup weekend, hackathon, game jam, SoT Hackfest or Design Day, this is similar! If you haven't heard of any of those things - THAT'S FINE - we'd love to chat more about why it's a fun & career-enhancing thing to do!

Web Challenge will focus on creativity, team work and working alongside mentors.

It is a great place to get some real-world experience, work alongside designers/UX students/grads to grow your communication & collaboration skills, deliver something to a client, work closely with mentors, and build something awesome for your portfolio. It’s heaps of fun!!

We have a "client" who are a government department. They have asked for our help to "re-imagine" one of their websites, to make it more user friendly and accessible and useful for users. So we want to bring together small teams of digital students & graduates to magic up some prototype solutions.

It's a big time commitment!! If you register, you are committing to attend the whole weekend, that's 9am-5pm Saturday, and 9am-6:30pm Sunday. We will feed you. There will be prizes/incentives. You will be mentored & supported & you will learn heaps. At the end of the weekend, you'll have experience, something for your portfolio, and LOTS of stuff to talk about during job interviews. Our "client" organisation is likely to be hiring interns this summer, so this would be a huge "foot in the door"!

If your plans change after you register for this event, it is really really important that you withdraw from the event. If it's within 48 hours of the start time, then you'll need to email us!!!

More details will follow about the project brief nearer the time!

If you have questions or would like to chat about the event, please get in touch via info@summeroftech.co.nz or send us a DM. We would LOVE to have you involved if you can spare a weekend as an investment in your future career!